requirements.txt
streamlit
pandas
openpyxl
plotly
xlsxwriter
mysql-connector-python
altair
numpy